refactor anat pipelines tests
